refactor: reorg and rename
