+++ src/08-recommendations/data/artwork-objects.ts
