+++ src/08-recommendations/data/artwork-texts.ts
